mpi first version
