+++ tasks/mpi/suvorov_d_sum_of_vector_elements/perf_tests/main.cpp
@@ -7,7 +7,7 @@
 #include "core/perf/include/perf.hpp"
 #include "mpi/example/include/ops_mpi.hpp"
 
-TEST(mpi_example_perf_test, test_pipeline_run) {
+TEST(suvorov_d_sum_of_vector_elements_mpi, test_pipeline_run) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;
   std::vector<int32_t> global_sum(1, 0);
@@ -23,7 +23,7 @@ TEST(mpi_example_perf_test, test_pipeline_run) {
     taskDataPar->outputs_count.emplace_back(global_sum.size());
   }
 
-  auto testMpiTaskParallel = std::make_shared<nesterov_a_test_task_mpi::TestMPITaskParallel>(taskDataPar, "+");
+  auto testMpiTaskParallel = std::make_shared<suvorov_d_sum_of_vector_elements::TestMPITaskParallel>(taskDataPar, "+");
   ASSERT_EQ(testMpiTaskParallel->validation(), true);
   testMpiTaskParallel->pre_processing();
   testMpiTaskParallel->run();
@@ -47,7 +47,7 @@ TEST(mpi_example_perf_test, test_pipeline_run) {
   }
 }
 
-TEST(mpi_example_perf_test, test_task_run) {
+TEST(suvorov_d_sum_of_vector_elements_mpi, test_task_run) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;
   std::vector<int32_t> global_sum(1, 0);
@@ -63,7 +63,7 @@ TEST(mpi_example_perf_test, test_task_run) {
     taskDataPar->outputs_count.emplace_back(global_sum.size());
   }
 
-  auto testMpiTaskParallel = std::make_shared<nesterov_a_test_task_mpi::TestMPITaskParallel>(taskDataPar, "+");
+  auto testMpiTaskParallel = std::make_shared<suvorov_d_sum_of_vector_elements::TestMPITaskParallel>(taskDataPar, "+");
   ASSERT_EQ(testMpiTaskParallel->validation(), true);
   testMpiTaskParallel->pre_processing();
   testMpiTaskParallel->run();
